* ns train validation fix

+++ teafacto/core/base.py
@@ -639,7 +639,7 @@ class NSTrainConfig():
         return t.train(*args, **kwargs)
 
     def validate_on(self, data, splits=1, random=False):
-        gold = np.ones((data[0].shape[0],), dtype="int32")
+        gold = np.ones((data[0].shape[0],), dtype="float32")
         self._trainerconfigstorer("validate_on", data, gold, splits=splits, random=random)
         self.linear_objective()
         return self
